+++ Tester_Publisher_and_Sub.py
@@ -1,11 +1,15 @@
 from MyMQTT import MyMQTT
 import time
 from random import randint
+import json
 
 class Publisher_Test():
     def __init__(self):
-        self.client = MyMQTT("aksjfgaj","test.mosquitto.org", 1883, None)
+        self.client = MyMQTT("aksjfgaj","test.mosquitto.org", 1883, self)
         self.client.start()
+        time.sleep(5)
+        # Subscribing
+        self.client.mySubscribe("group01/IoTproject/PatientAlarm/Antonio")
     
     def send(self):
         b=randint(2,3)
@@ -16,7 +20,7 @@ class Publisher_Test():
         Pi=[randint(4,100) for i in range(9)]
         print(f"Stringa PI generata: {Pi}")
         s=[randint(94,100) for i in range(9)]
-        #print(f"Stringa sat generata: {s}")
+        print(f"Stringa sat generata: {s}")
         p=[randint(65,145) for i in range(9)]
         print(f"Stringa pulse generata: {p}")
 
@@ -24,8 +28,16 @@ class Publisher_Test():
         msg_temp={"bn" : "ID_sat", "bt" : time.time(), "e" : [{ "n":"battery", "u" : "V", "v" :b }, { "n":"temperature", "u" : "cel", "v" :t }]}
 
         self.client.myPublish("group01/IoTProject/Patient/Pulsation/Antonio",msg_sat)
-        time.sleep(20)
+        time.sleep(10)
         self.client.myPublish("group01/IoTProject/Patient/Temperature/Antonio",msg_temp)
+        print("All published\n")
+
+
+    def notify(self,topic,msg): # Metodo che analizza i dati arrivati e li stampa
+        msg=json.loads(msg)
+        ID_P=topic.split("/")[-1]
+        print(f"Arrivato Paziente {ID_P}")
+        print(msg)
 
 if __name__=="__main__":
   P=Publisher_Test()
